Homework 28. Refactoring (added modules).
