+++ hw26/css/main.css
@@ -34,6 +34,7 @@ body {
     padding: 10px;
 
     margin-bottom: 10px;
+    align-self: start;
 }
