+++ hw26/index.html
@@ -117,7 +117,7 @@
     <div class="overlay"></div>
 </div>
 
-<script src="./scripts/main.js"></script>
+<script src="./scripts/main.js" type="module"></script>
 </body>
 
 </html>
