+++ hw26/scripts/aside.js
@@ -1,0 +1,12 @@
+// aside
+const aside = document.getElementById("aside");
+aside.addEventListener("click", removeOrderInfoBlock);
+
+function removeOrderInfoBlock(e) {
+    const clickedEl = e.target;
+
+    if (clickedEl.className === "close-btn") {
+        const order = clickedEl.closest(".order");
+        order.remove();
+    }
+};
